+++ wtfix/message/tests/test_fieldset.py
@@ -1,11 +1,9 @@
-import json
 import uuid
 from datetime import datetime
 
 import pytest
 
 from wtfix.conf import settings
-from wtfix.core.json import FieldSetEncoder, FieldSetDecoder
 from wtfix.message.message import generic_message_factory
 from wtfix.protocol.common import Tag, MsgType
 from ..field import Field
@@ -276,20 +274,6 @@ class TestListFieldSet:
         fs = ListFieldSet((34, "a"), (35, "bb"), (1, "ccc"))
         assert str(fs) == "[MsgSeqNum (34):a | MsgType (35):bb | Account (1):ccc]"
 
-    def test_serialization_is_idempotent(self, nested_parties_group):
-        fs = ListFieldSet(
-            (1, "a"),
-            (2, "b"),
-            nested_parties_group.identifier,
-            *nested_parties_group.fields,
-            (3, "c"),
-        )
-
-        encoded = json.dumps(fs, cls=FieldSetEncoder)
-        decoded = json.loads(encoded, cls=FieldSetDecoder)
-
-        assert fs == decoded
-
 
 class TestOrderedDictFieldSet:
     def test_parse_fields_duplicate_tags_raises_exception(self):
@@ -508,20 +492,6 @@ class TestOrderedDictFieldSet:
                 9960,
             ]
         )
-
-    def test_serialization_is_idempotent(self, nested_parties_group):
-        fs = OrderedDictFieldSet(
-            (1, "a"),
-            (2, "b"),
-            nested_parties_group.identifier,
-            *nested_parties_group.fields,
-            (3, "c"),
-            group_templates={539: [524, 525, 538, 804], 804: [545, 805]},
-        )
-        encoded = json.dumps(fs, cls=FieldSetEncoder)
-        decoded = json.loads(encoded, cls=FieldSetDecoder)
-
-        assert fs == decoded
 
 
 class TestGroup: